json: do not enforce snake_case
Enforcing snake_case is silly, if you develop a client and the backend expect
your payload to be in cameCase or PascalCase or any cases this tool will
bascially not fit your need.

It is better to rely on the properties keys defined in the spec yaml fine such
that you could adapt to what your backend is expecting.

Signed-off-by: Lionel H <me@nullbyte.be>